Updates the set list page viewmodel and navgraph
