+++ app/src/main/java/org/wspcgir/strong_giraffe/destinations/edit_variation/Navigation.kt
@@ -19,7 +19,6 @@ import androidx.navigation.toRoute
 import kotlinx.coroutines.launch
 import kotlinx.serialization.Serializable
 import org.wspcgir.strong_giraffe.destinations.EditLocation
-import org.wspcgir.strong_giraffe.destinations.edit_set.EditSet
 import org.wspcgir.strong_giraffe.model.Location
 import org.wspcgir.strong_giraffe.repository.AppRepository
 import org.wspcgir.strong_giraffe.views.SelectionPage
